+++ app/src/main/java/com/jgbravo/tvbillboard/data/remote/responses/models/MetadataModel.kt
@@ -8,5 +8,5 @@ data class MetadataModel(
     @SerializedName("fullLength")
     val fullLength : Int?,
     @SerializedName("timestamp")
-    val timestamp : Int?
+    val timestamp : Long?
 )
